+++ src/index.js
@@ -1,23 +1,23 @@
-var download = require('./download');
+var downloadFileFromBucket = require('./downloadFileFromBucket');
 var downloadClamscanDbFiles = require('./downloadClamscanDbFiles');
-var downloadUrlToFile = require('./util/downloadUrlToFile');
+var downloadFileFromUrl = require('./util/downloadFileFromUrl');
 var getClamscan = require('./getClamscan');
 var getS3 = require('./getS3');
 var getSns = require('./getSns');
 var handler = require('./handler');
 var initClamscan = require('./initClamscan');
-var scan = require('./scan');
-var sns = require('./sns');
+var scanFile = require('./scanFile');
+var notifySns = require('./notifySns');
 
 module.exports = {
-  download: download,
   downloadClamscanDbFiles: downloadClamscanDbFiles,
-  downloadUrlToFile: downloadUrlToFile,
+  downloadFileFromBucket: downloadFileFromBucket,
+  downloadFileFromUrl: downloadFileFromUrl,
   getClamscan: getClamscan,
   getS3: getS3,
   getSns: getSns,
   handler: handler,
   initClamscan: initClamscan,
-  scan: scan,
-  sns: sns
+  notifySns: notifySns,
+  scanFile: scanFile
 };
