+++ src/scanFile.js
